% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{add_perc_host_metadata}
\alias{add_perc_host_metadata}
\title{Add percent host column to metadata}
\usage{
add_perc_host_metadata(metadata, sampleID, report, host_col = "Percent host")
}
\arguments{
\item{metadata}{A data frame containing metadata for the samples,
where rows are samples}

\item{sampleID}{A string representing the column name in `metadata`
that contains the sample IDs.
This column is used to merge the metadata with the report file table.}

\item{report}{A dataframe that contains the percent host
column `host_col` and the `sampleID` column.}

\item{host_col}{A string representing the column name in `report`
that contains the percent host per sample.
This column will be added to the `metadata`.}
}
\value{
A dataframe containing the metadata sample information with the
host percent column appended at the end.
}
\description{
Add percent host column to metadata
}
\examples{
metadata <- data.frame(SampleID=paste0("sample_", seq_len(4)),
                          phenotype=c(0,0,1,1),
                          participant=c(0,1,0,1),
                          timepoint=c(0,0,1,1))
report <- data.frame(SampleID=paste0("sample_", seq_len(4)),
                     percent_host=c(0.9, 0.8, 0.5, 0.9))
new_metadata <- add_perc_host_metadata(metadata, "SampleID",
                                       report, "percent_host")

}
