% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoMiCorr.R
\name{check_duplicated_colnames}
\alias{check_duplicated_colnames}
\title{Check for duplicated values between 2 sets of column names}
\usage{
check_duplicated_colnames(colnames1, colnames2)
}
\arguments{
\item{colnames1}{vector of strings with column names from the first dataset}

\item{colnames2}{vector of strings with column names from the second dataset}
}
\value{
Nothing
}
\description{
Ensures there are no intersecting values between two vectors
}
