% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{check_for_ones}
\alias{check_for_ones}
\title{Check for 1 in feature.table}
\usage{
check_for_ones(feature.table)
}
\arguments{
\item{feature.table}{A dataframe (presumably metatranscriptome relative
expression levels), where rows are samples,
and columns are genes/features. Row names should be sample IDs.
Data values should be relative abundances between 0 and 1.}
}
\value{
Nothing
}
\description{
The zero-inflated beta regression can't handle values of 1.
This checks for them and raises an error if they exist.
}
\examples{
data <- data.frame(a=c(0.1, 0.2, 0.3, 0.4),
                 b=c(0, 0, 1, 1))

# This would raise an error about rows 3 and 4
try(check_for_ones(data))

}
