% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{.add_dna_to_formula}
\alias{.add_dna_to_formula}
\title{Add matching dna column to formula or fixed vars}
\usage{
.add_dna_to_formula(data, col, formula, fixed.vars, reg.method)
}
\arguments{
\item{data}{A data frame containing the merged feature table and metadata,
ready for regression.}

\item{col}{A string representing the name of the feature (column)
in `data` to be analyzed.}

\item{formula}{A string representing the formula for the regression model.}

\item{fixed.vars}{A vector of strings representing the fixed effect
variables to include in the regression.}

\item{reg.method}{A string indicating the regression method to be used.
Options include "zibr", "gamlss", "lm", and "lmer".}
}
\value{
list of updated `formula` and `fixed.vars` with the matching dna
column added to them
}
\description{
This function adds the matching dna feature of a `col` to a formula or
fixed vars
}
\keyword{internal}
