% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{.add_host_to_formula_metadata}
\alias{.add_host_to_formula_metadata}
\title{Add percent host column to formula}
\usage{
.add_host_to_formula_metadata(
  metadata,
  formula,
  sampleID,
  report = NULL,
  host_col = NULL
)
}
\arguments{
\item{metadata}{A data frame containing the merged feature table and metadata,
ready for regression.}

\item{formula}{A string representing the formula for the regression model.}

\item{sampleID}{A string representing the column name in `metadata`
that contains the sample IDs.}

\item{report}{A dataframe that contains the percent host
column `host_col` and the `sampleID` column.}

\item{host_col}{A string representing the name of the percent host column
within the `data` data.frame.}
}
\value{
list of updated `formula` and `fixed.vars` with the
`host_col`added to them
}
\description{
Add percent host column to formula
}
\keyword{internal}
