% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMs.R
\name{.add_K_line_to_matrix}
\alias{.add_K_line_to_matrix}
\title{Adds KEGG line to the GMM matrix}
\usage{
.add_K_line_to_matrix(GMM.matrix, line, mod.name, mod.number)
}
\arguments{
\item{GMM.matrix}{The matrix with GMM data}

\item{line}{A string containing the line
to be parsed and added to the GMM matrix}

\item{mod.name}{A string containing the name of the module
for which current line is being parsed}

\item{mod.number}{A string containing the ID of the module
for which current line is being parsed}
}
\value{
The updated matrix of GMMs, with the new line parsed and added
}
\description{
Parser for each line of the GMM file
}
\keyword{internal}
