% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoMiCorr.R
\name{.adjust_p_values}
\alias{.adjust_p_values}
\title{Adjust P-values for Multiple Comparisons (internal)}
\usage{
.adjust_p_values(mod.summaries, reg.method, zero_prop_from_formula, padj)
}
\arguments{
\item{mod.summaries}{A data frame containing model summary results,
including p-values.}

\item{reg.method}{A string indicating the regression method used.
Options include "zibr", "gamlss", "lm", and "lmer".}

\item{zero_prop_from_formula}{A boolean indicating whether zeroes
were modeled using the provided formula (relevant for ZIBR regression).}

\item{padj}{A string specifying the multiple testing correction method.
Supported methods can be checked with `p.adjust.methods`.}
}
\value{
A modified version of `mod.summaries` with an additional column
`q`, containing the adjusted p-values.
}
\description{
This function applies multiple testing correction
to the p-values from regression model summaries. The method used for
adjustment is specified by the user.
}
\examples{
# Example model summary
mod.summaries <- data.frame(term = c("(Intercept)", "GeneX"),
                            parameter = c("beta", "beta"),
                            p.value = c(0.01, 0.002),
                            joint.p = c(0.02, 0.005))

# Adjust p-values using FDR correction
mod.summaries <- HoMiStats:::.adjust_p_values(mod.summaries, "zibr",
                                 zero_prop_from_formula = TRUE,
                                 padj = "fdr")

}
\keyword{internal}
