% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoMiCorr.R
\name{.generate_feature_combinations}
\alias{.generate_feature_combinations}
\title{Generate Feature Combinations for Regression Analysis (internal)}
\usage{
.generate_feature_combinations(mtx, host)
}
\arguments{
\item{mtx}{A data frame containing microbial gene expression data,
where rows represent samples and columns represent genes.}

\item{host}{A data frame containing host gene expression data,
where rows represent samples and columns represent genes.}
}
\value{
A list of feature index pairs, representing all unique
combinations of microbial and host genes.
}
\description{
This function generates all possible pairwise combinations
of microbial and host gene features.
}
\examples{
# Simulated data
mtx <- data.frame(a1 = c(0.1, 0.0, 0.0, 0.0),
                  b1 = c(0.5, 0.5, 0.5, 0.4))
host <- data.frame(a2 = c(0.5, 0.0, 0.0, 0.0),
                   b2 = c(0.5, 0.5, 0.5, 0.4))

# Generate feature combinations
feature_combos <- HoMiStats:::.generate_feature_combinations(mtx, host)

}
\keyword{internalâ }
