% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{.prepare_data_mtxDE}
\alias{.prepare_data_mtxDE}
\title{Prepare Data for Differential Expression Analysis (internal)}
\usage{
.prepare_data_mtxDE(feature.table, metadata, formula, sampleID, zibr_time_ind)
}
\arguments{
\item{feature.table}{A data frame where rows are samples and
columns are features (e.g., genes).
Row names should correspond to sample IDs.}

\item{metadata}{A data frame containing metadata for the samples,
where rows are samples
and columns are metadata variables (e.g., phenotype, timepoint).}

\item{formula}{A string representing the right-hand side of the
regression formula to be used
in the analysis. All variables in this formula should be numeric.}

\item{sampleID}{A string representing the column name in `metadata`
that contains the sample IDs.
This column is used to merge the metadata with the feature table.}

\item{zibr_time_ind}{A string representing the time column in `metadata` for
Zero-Inflated Beta Regression (ZIBR).
This is used only when the regression method involves ZIBR.}
}
\value{
A merged data frame containing the feature table and metadata,
ready for regression analysis.
}
\description{
This function prepares the data by merging the feature table
with metadata based on the sample IDs.
It ensures that the sample IDs in the feature table match the
sample IDs in the metadata and validates the input for any inconsistencies.
}
\keyword{internal}
