% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{.run_regressions_mtxDE}
\alias{.run_regressions_mtxDE}
\title{Run Multiple Regressions for Features in Parallel (internal)}
\usage{
.run_regressions_mtxDE(
  data,
  reg.method,
  feature.table,
  formula,
  zero_prop_from_formula,
  zibr_time_ind,
  ncores,
  show_progress
)
}
\arguments{
\item{data}{A data frame containing the merged feature table
and metadata, prepared for regression.}

\item{reg.method}{A string indicating the regression method to be used.
Options include
"zibr", "gamlss", "lm", and "lmer".}

\item{feature.table}{A data frame where rows represent samples and
columns represent features.}

\item{formula}{A string representing the formula for the regression model.}

\item{zero_prop_from_formula}{A boolean indicating whether zeroes
should be modeled using the provided formula (for ZIBR).}

\item{zibr_time_ind}{A string representing the time column in `data`
for ZIBR, or NULL if not applicable.}

\item{ncores}{An integer specifying the number of cores to use
for parallel processing.}

\item{show_progress}{A boolean indicating whether a progress bar
should be shown during parallel computation.}
}
\value{
A data frame containing the regression summaries for all features.
The summaries include terms,
estimates, standard errors, statistics, p-values, and feature information.
}
\description{
This function runs regressions for each feature (column)
in the `feature.table` using the specified regression method.
The function handles different regression methods such as
Zero-Inflated Beta Regression (ZIBR),
Generalized Additive Models for Location Scale and Shape (GAMLSS),
linear regression (LM), and linear mixed effects regression (LME).
}
\keyword{internal}
