% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{.run_single_regssion_mtxDE}
\alias{.run_single_regssion_mtxDE}
\title{Run a Single Regression for a Feature (internal)}
\usage{
.run_single_regssion_mtxDE(
  data,
  reg.method,
  col,
  formula,
  fixed.vars = NULL,
  random.effects.vars = NULL,
  zibr_time_ind = NULL,
  zero_prop_from_formula = NULL
)
}
\arguments{
\item{data}{A data frame containing the merged feature table and metadata,
ready for regression.}

\item{reg.method}{A string indicating the regression method to be used.
Options include "zibr", "gamlss", "lm", and "lmer".}

\item{col}{A string representing the name of the feature (column)
in `data` to be analyzed.}

\item{formula}{A string representing the formula for the regression model.}

\item{fixed.vars}{A vector of strings representing the fixed effect
variables to include in the regression.}

\item{random.effects.vars}{A vector of strings representing the
random effect variables to include in the regression.}

\item{zibr_time_ind}{A string representing the time column in `data`
for ZIBR, or NULL if not applicable.}

\item{zero_prop_from_formula}{A boolean indicating whether zeroes should
be modeled using the provided formula (for ZIBR).}
}
\value{
A data frame containing the summary of the regression results
for the specified feature.
The summary includes terms, estimates, standard errors, statistics,
p-values, and feature information.
}
\description{
This function runs a single regression model for
a given feature using the specified regression method.
It handles different types of regression methods, including
Zero-Inflated Beta Regression (ZIBR),
Generalized Additive Models for Location Scale and Shape
(GAMLSS),linear regression (LM), and linear mixed effects regression (LME).
}
\keyword{internal}
