% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoMiCorr.R
\name{.stop_parallel_processing}
\alias{.stop_parallel_processing}
\title{Stop Parallel Processing and Clean Up (internal)}
\usage{
.stop_parallel_processing(cl, show_progress, pb)
}
\arguments{
\item{cl}{A parallel cluster object created by `parallel::makeCluster`.}

\item{show_progress}{A boolean indicating whether a progress bar was shown.
If TRUE, the function attempts to close the progress bar.}
}
\value{
This function does not return a value. It stops the cluster
and cleans up resources.
}
\description{
This function stops the parallel processing cluster
and closes the progress bar if `show_progress` is enabled.
}
\examples{
# Example usage (not run):
# cl <- parallel::makeCluster(2)
# HoMiStats:::.stop_parallel_processing(cl, show_progress = FALSE)

}
\keyword{internal}
