% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted.R
\name{features_from_gmm_df}
\alias{features_from_gmm_df}
\title{Pull mtx features for a gut metabolic module of interest}
\usage{
features_from_gmm_df(GMM, GMM.kos.df, mtx.feature.names)
}
\arguments{
\item{GMM}{The name of the gut metabolic module of interest}

\item{GMM.kos.df}{A data.frame with columns "Module" and "KEGG" with the KOs corresponding to each gut metabolic module. Can be generated using get_GMM_matrix()}

\item{mtx}{A vector with all of the metatranscriptome feature names, with KOs in them}
}
\value{
a vector of feature names corresponding to the GMM of interest
}
\description{
Looks for all mtx features corresponding to the gut metabolic module of interest
}
