% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_sparse_features}
\alias{filter_sparse_features}
\title{Filter sparse features from relative abundance data}
\usage{
filter_sparse_features(df, threshold, zero_threshold = 0)
}
\arguments{
\item{df}{A dataframe containing relative abundance data, where rows are samples and columns are features (genes).}

\item{threshold}{The threshold for filtering sparse features. Any features more sparse than this will be removed.}

\item{zero_threshold}{The threshold below which a sample is considered not present. Defaults to zero}
}
\value{
A dataframe with sparse features filtered out.
}
\description{
Filters features with sparsity above a certain threshold.
}
\examples{
df <- data.frame(
  feature1 = c(0.0, 0.2, 0.3, 0.4),
  feature2 = c(0.2, 0.3, 0.4, 0.5),
  feature3 = c(0.3, 0.4, 0.5, 0.6)
)
filtered_df <- filter_low_abundance(df, 0.1)
print(filtered_df)
}
