% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{filter_tables_by_shared_columns}
\alias{filter_tables_by_shared_columns}
\title{Filter two tables by shared columns}
\usage{
filter_tables_by_shared_columns(table1, table2, table1_name, table2_name)
}
\arguments{
\item{table1}{A data frame to be filtered.}

\item{table2}{A second data frame to be filtered.}

\item{table1_name}{A character string specifying the name of `table1`.}

\item{table2_name}{A character string specifying the name of `table2`}
}
\value{
A list containing:
\describe{
  \item{`table1_name`}{The filtered version of `table1`,
  keeping only the shared columns.}
  \item{`table2_name`}{The filtered version of `table2`,
  keeping only the shared columns.}
}
}
\description{
This function filters two data frames
to only retain shared columns.
}
\examples{
feature.table <- data.frame(a = c(0.1, 0.2),
                            b = c(0.3, 0.4),
                            c = c(0.5, 0.6))
dna.table <- data.frame(a = c(0.7, 0.8),
                        b = c(0.9, 1.0))
result <- filter_tables_by_shared_columns(feature.table, dna.table,
                                          "feature.table", "dna.table")
filtered_feature <- result$feature.table
filtered_dna <- result$dna.table

}
