% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{filter_undetected}
\alias{filter_undetected}
\title{Filter undetected features from the feature.table}
\usage{
filter_undetected(feature.table)
}
\arguments{
\item{feature.table}{A dataframe, where rows are samples,
and columns are genes/features. Row names should be sample IDs.}
}
\value{
A dataframe, with undetected features removed
(where rows are samples, and columns are genes/features).
}
\description{
The beta regression methods don't deal well
with features that are entirely undetected.
This filters them from the feature table.
}
\examples{
good.data <- data.frame(a=c(0.1, 0.2, 0.3, 0.4),
                   b=c(0, 0, 1, 1))
bad.data <- data.frame(a=c(0.1, 0.2, 0.3, 0.4),
                   b=c(0, 0, 1, 1),
                   undetected_feature=c(0, 0, 0, 0))

# This will not raise any warning
filter_undetected(good.data)

# This would raise a warning
try(filter_undetected(bad.data))

}
