% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMs.R
\name{get_GMM_matrix}
\alias{get_GMM_matrix}
\title{Gets the GMM matrix in a tabular format}
\usage{
get_GMM_matrix(
  filepath = "GMMs.txt",
  version = "v1.07",
  quietly_download = TRUE,
  cleanup = TRUE
)
}
\arguments{
\item{filepath}{The filepath at which the GMMs file should be downloaded or found.}

\item{version}{The version of GMMs to be used. As of April '24, the newest version is v1.07. This is used in pulling the data from the Raes lab GMMs github, so old versions that are not on the master branch may cause the filepath to not work.}

\item{cleanup}{A boolean denoting if the filepath should be removed at the end of this function.}

\item{quietly}{A boolean denoting if the GMMs file should be downloaded quietly}
}
\value{
A tabular version of the GMMs as a dataframe
}
\description{
Download the GMM data (if needed) and parses this file to return a table with the GMMs and corresponding KOs
}
