% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted.R
\name{get_go_term_human_genes}
\alias{get_go_term_human_genes}
\title{Pull all human genes corresponding to a GO term of interest}
\usage{
get_go_term_human_genes(go.term)
}
\arguments{
\item{go.term}{A string version of a GO term, such as "GO:0042110"}
}
\value{
a vector of gene names corresponding to the GO term of interest
}
\description{
Looks for all genes corresponding to the GO term of interest
}
