% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted.R
\name{go_targeted_diffex}
\alias{go_targeted_diffex}
\title{Run targeted differential expression analysis in DESeq2 for a subset of genes}
\usage{
go_targeted_diffex(
  targeted.genes,
  host.genes,
  microbial.gene,
  microbial.genes,
  verbose = T
)
}
\arguments{
\item{targeted.genes}{A vector of host gene names}

\item{host.genes}{A matrix or data.frame with the host gene count data}

\item{microbial.gene}{The name of the microbial gene to be used}

\item{microbial.genes}{A matrix or data.frame with the microbial gene count data}
}
\value{
A dataframe of DESeq2 results
}
\description{
Uses the vector of genes provided to subset the data then run DESeq2
}
