% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted.R
\name{GO_targeted_for_each_KO_within_GMM}
\alias{GO_targeted_for_each_KO_within_GMM}
\title{Run differential expression testing on a targeted subset of genes}
\usage{
GO_targeted_for_each_KO_within_GMM(
  go.terms,
  host.genes,
  mtx,
  mtx.features,
  padj = "fdr",
  verbose = FALSE
)
}
\arguments{
\item{go.terms}{A string vector with the GO terms of interest}

\item{host.genes}{A matrix or data.frame with the host gene count data}

\item{mtx}{A data.frame with the normalized microbial gene count data}

\item{mtx.features}{A vector of features to use from the microbial gene counts}

\item{padj}{A string denoting the p value adustment method. Options can be checked using 'p.adjust.methods'}
}
\value{
a data.frame with DESeq results for each mtx feature
}
\description{
Run differential expression testing on a targeted subset of genes
}
