% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{map_zibr_termnames}
\alias{map_zibr_termnames}
\title{Map ZIBR term names}
\usage{
map_zibr_termnames(data, real.names)
}
\arguments{
\item{data}{a character vector of the ZIBR terms (e.g. c("intercept", "var1", "var2"))}

\item{real.names}{names of variables in the order they were passed to ZIBR. Do not include intercept}
}
\value{
a character vector with the mapped variable names
}
\description{
ZIBR outputs terms as "var1", "var2", etc in the results. This maps these to the original variable names
}
\examples{
data <- data.frame(y=c(0.1, 0, 0.3, 0.4),
                   treated=c(0, 0, 0, 1),
                   participant=c(1, 2, 1, 2),
                   timepoint=c(0, 0, 1, 1))

# This runs ZIBR where treatment impacts both whether a feature might be zero
# and the abundance of that feature (and random intercepts for participant)
mod <- run_single_beta_reg_zibr(logistic_cov="treated",
                                beta_cov="treated",
                                Y="y",
                                subject_ind="participant",
                                time_ind="timepoint",
                                data=data)
tidied.mod <- tidy_zibr_results(mod)
tidied.mod$term <- map_zibr_termnames(tidied.mod$term, c("treated"))

}
