% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMs.R
\name{pull_GMMs_file}
\alias{pull_GMMs_file}
\title{Download GMMs file}
\usage{
pull_GMMs_file(filepath, version = "v1.07", quietly = TRUE)
}
\arguments{
\item{filepath}{The filepath where the GMMs file should be downloaded}

\item{version}{The version of GMMs to be used.
As of April '24, the newest version is v1.07.
This is used in pulling the data from the Raes lab GMMs github,
so old versions that are not on the master branch
may cause the filepath to not work.}

\item{quietly}{A boolean denoting if the GMMs file
should be downloaded quietly}
}
\value{
Nothing
}
\description{
Pulls the Raes lab gut metabolic modules file (v1.0.7).
This will not overwrite an existing file.
}
\examples{
pull_GMMs_file("loc_to_download.txt")

# cleanup
file.remove("loc_to_download.txt")
}
