% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{run_single_beta_reg_gamlss}
\alias{run_single_beta_reg_gamlss}
\title{Run a single zero-inflated beta regression using gamlss}
\usage{
run_single_beta_reg_gamlss(
  formula,
  data,
  controller = gamlss.control(trace = FALSE)
)
}
\arguments{
\item{formula}{the full formula to be used in the regression}

\item{data}{A dataframe to be used in the regression}
}
\value{
the resulting model
}
\description{
runs a zero-inflated beta distribution GAM using gamlss
}
