% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{run_single_beta_regression}
\alias{run_single_beta_regression}
\title{Run a single zero-inflatd beta regression using gamlss}
\usage{
run_single_beta_regression(formula, data)
}
\arguments{
\item{formula}{the full formula to be used in the regression}

\item{data}{A dataframe to be used in the regression}
}
\value{
the resulting model
}
\description{
runs a zero-inflated beta distribution GAM using gamlss
}
