% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{run_single_lmer}
\alias{run_single_lmer}
\title{Run a single linear mixed effects regression}
\usage{
run_single_lmer(formula, data)
}
\arguments{
\item{formula}{the full formula to be used in the regression}

\item{data}{A dataframe to be used in the regression}
}
\value{
the resulting model
}
\description{
runs a linear mixed regression using lme4
}
\examples{
data <- data.frame(a=c(0.1, 0.2, 0.3, 0.4),
                   b=c(0, 0, 1, 1),
                   participant=c(0, 1, 0, 1))

run_single_lmer("a ~ b + (1|participant)",
              data=data)

}
