% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{tidy_zibr_results}
\alias{tidy_zibr_results}
\title{Tidy ZIBR output}
\usage{
tidy_zibr_results(mod)
}
\arguments{
\item{mod}{A ZIBR model results object.}
}
\value{
a dataframe with the tidy model results
}
\description{
ZIBR returns a list with lots of different kinds of entrie
}
