% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformers.R
\name{transform_feature_table}
\alias{transform_feature_table}
\title{Transform feature table for analysis}
\usage{
transform_feature_table(feature.table, transformation)
}
\arguments{
\item{feature.table}{A dataframe, where rows are samples, and columns are genes/features. Row names should be sample IDs.}

\item{transformation}{Method for transforming the feature.table. String options include NOT YET IMPLEMENTED, or transformation can be a function used for transforming data}
}
\value{
The transformed table
}
\description{
transforms feature table for differential expression or co-occurrence analysis
}
