% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoMiCorr.R
\name{check_duplicated_colnames}
\alias{check_duplicated_colnames}
\title{Check for duplicated values between 2 sets of column names}
\usage{
check_duplicated_colnames(colnames1, colnames2)
}
\arguments{
\item{colnames1}{vector of strings with column names from the first dataset}

\item{colnames2}{vector of strings with column names from the second dataset}
}
\value{
Nothing
}
\description{
Ensures there are no intersecting values between two vectors
}
\examples{
# Simulate data
mtx <- data.frame(a1=c(0.1, 0.0, 0.0, 0.0),
                 b1=c(0.5, 0.5, 0.5, 0.4),
                 c1=c(0.4, 0.5, 0.0, 0.0),
                 d1=c(0.0, 0.0, 0.5, 0.6))
host <- data.frame(a2=c(0.5, 0.0, 0.0, 0.0),
                 b2=c(0.5, 0.5, 0.5, 0.4),
                 c2=c(0.4, 0.5, 0.0, 0.0),
                 d2=c(0.0, 0.0, 0.5, 0.6))

# This should not raise an error
check_duplicated_colnames(colnames(mtx), colnames(host))
# This should raise an error
try(check_duplicated_colnames(colnames(mtx), c("a1")))

}
