% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{check_for_ones}
\alias{check_for_ones}
\title{Check for 1 in feature.table}
\usage{
check_for_ones(feature.table)
}
\arguments{
\item{feature.table}{A dataframe, where rows are samples, and columns are genes/features. Row names should be sample IDs.}
}
\value{
Nothing
}
\description{
The zero-inflated beta regression can't handle values of 1. This checks for them and raises an error if they exist.
}
