% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted.R
\name{get_go_term_human_genes}
\alias{get_go_term_human_genes}
\title{Retrieve human genes associated with a GO term}
\usage{
get_go_term_human_genes(go.term)
}
\arguments{
\item{go.term}{A character string representing a GO term, such as "GO:0042110".}
}
\value{
A character vector of human gene symbols associated with the GO term.
}
\description{
Fetches all human genes that are annotated with a given Gene Ontology (GO) term.
}
\examples{
genes <- get_go_term_human_genes("GO:0000009")

}
