% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{get_random_fx}
\alias{get_random_fx}
\title{Get random effects}
\usage{
get_random_fx(form)
}
\arguments{
\item{form}{A formula with random effects.}
}
\value{
a character vector with the random effects variable names, or NULL if there are no random effects.
}
\description{
Sometimes you want the name of random effects variables from a formula. This function does that.
}
