% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{map_zibr_termnames}
\alias{map_zibr_termnames}
\title{Map ZIBR term names}
\usage{
map_zibr_termnames(data, real.names)
}
\arguments{
\item{data}{a character vector of the ZIBR terms (e.g. c("intercept", "var1", "var2"))}

\item{real.names}{names of variables in the order they were passed to ZIBR. Do not include intercept}
}
\value{
a character vector with the mapped variable names
}
\description{
ZIBR outputs terms as "var1", "var2", etc in the results. This maps these to the original variable names
}
