% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMMs.R
\name{pull_GMMs_file}
\alias{pull_GMMs_file}
\title{Download GMMs file}
\usage{
pull_GMMs_file(filepath, version = "v1.07", quietly = TRUE)
}
\arguments{
\item{filepath}{The filepath where the GMMs file should be downloaded}
}
\value{
Nothing
}
\description{
Pulls the Raes lab gut metabolic modules file (v1.0.7). Will not overwrite an existing file.
}
