% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HoMiCorr.R
\name{run_HoMiCorr}
\alias{run_HoMiCorr}
\title{Create a correlation network between host and microbial data}
\usage{
run_HoMiCorr(
  mtx,
  host,
  covariates = NULL,
  metadata = NULL,
  sampleID = NULL,
  reg.method = "zibr",
  padj = "fdr",
  zero_prop_from_formula = TRUE,
  zibr_time_ind = NULL,
  ncores = 1,
  show_progress = TRUE
)
}
\arguments{
\item{mtx}{A dataframe of metatranscriptome gene counts,
where rows are samples, and columns are metatranscriptome genes.
Row names should be sample IDs.}

\item{host}{A dataframe of tpm-normalized host gene counts,
where rows are samples, and columns are host genes.
Row names should be sample IDs.}

\item{covariates}{A string formula with covariates to be included
in each regression.
All variables in this formula should be of a numeric type
(factors should be dummy coded, such that the reference value is 0).}

\item{metadata}{The metadata dataframe,
where rows are samples and columns are features.
Only needed if passing covariates.}

\item{sampleID}{A string denoting name of the column in metadata with
sample IDs,
which should be used to merge metadata with the feature table's rownames}

\item{reg.method}{A string denoting the method to use for regression.
Options include "zibr" (Zero-inflated beta regression with random effects),
"gamlss" (Zero-inflated beta regression implemented via GAMLSS),
"lm" (linear regression),
and "lmer" (linear mixed effects regression
implemented via lme4 and lmerTest).}

\item{padj}{A string denoting the p value adustment method.
Options can be checked using 'p.adjust.methods'}

\item{zero_prop_from_formula}{In ZIBR zero-inflated beta regression,
should the zeroes be modeled with the provided formula? Default is TRUE.}

\item{ncores}{An integer denoting the number of cores to use if
running in parallel. Defaults to 1 (not parallelized).}

\item{show_progress}{A boolean denoting if a progress bar should be shown.}

\item{zibr_zibr_time_ind}{A string denoting the name of the time column
for ZIBR.
Defaults to NULL, which is implemented as a constant time value in ZIBR
to not fit a time effect.
This argument does nothing if reg.method is not "zibr".}
}
\value{
A dataframe with the model summaries as an adjacency list
}
\description{
Runs regressions between all features
and returns the model outputs
}
\examples{
# Simulate data
mtx <- data.frame(a1=c(0.1, 0.0, 0.0, 0.0),
                  b1=c(0.5, 0.5, 0.5, 0.4),
                  c1=c(0.4, 0.5, 0.0, 0.0),
                  d1=c(0.0, 0.0, 0.5, 0.6))
host <- data.frame(a2=c(0.5, 0.0, 0.0, 0.0),
                   b2=c(0.5, 0.5, 0.5, 0.4),
                   c2=c(0.4, 0.5, 0.0, 0.0),
                   d2=c(0.0, 0.0, 0.5, 0.6))
row.names(mtx) <- paste0("sample_", seq_len(4))
row.names(host) <- paste0("sample_", seq_len(4))
metadata <- data.frame(SampleID=paste0("sample_", seq_len(4)),
                       phenotype=c(0,0,1,1),
                       participant=c(0,1,0,1),
                       timepoint=c(0,0,1,1))

# Run HoMiCorr using a zero inflated beta regression
# with random effects
homicorr.res.zibr <- run_HoMiCorr(mtx, host,
                                reg.method="zibr",
                                covariates="(1|participant)",
                                metadata=metadata,
                                sampleID="SampleID",
                                zibr_time_ind="timepoint")

# Run HoMiCorr using a (faster) zero inflated beta regression
# without random effects
homicorr.res.gamlss <- run_HoMiCorr(mtx, host,
                                reg.method="gamlss",
                                show_progress=FALSE)


}
