% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{run_mtxDE}
\alias{run_mtxDE}
\title{Run differential expression analysis for metatranscriptomics data}
\usage{
run_mtxDE(formula, feature.table, metadata, sampleID, padj = "fdr")
}
\arguments{
\item{formula}{the right hand side of the regression formula to be used for each feature}

\item{feature.table}{A dataframe, where rows are samples, and columns are genes/features. Row names should be sample IDs.}

\item{metadata}{The metadata dataframe, where rows are samples and columns are features}

\item{sampleID}{The name of the column in metadata with sample IDs, which should be used to merge metadata with the feature table's rownames}

\item{padj}{p value adustment method. Options can be checked using 'p.adjust.methods'}
}
\value{
a dataframe with differential expression results
}
\description{
Regresses each feature using the formula provided and returns the statistical summary for each feature
}
