% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtxDE.R
\name{run_single_lm}
\alias{run_single_lm}
\title{Run a single linear regression}
\usage{
run_single_lm(formula, data)
}
\arguments{
\item{formula}{the full formula to be used in the regression}

\item{data}{A dataframe to be used in the regression}
}
\value{
the resulting model
}
\description{
runs a simple linear regression
}
\examples{
data <- data.frame(a=c(0.1, 0.2, 0.3, 0.4),
                   b=c(0, 0, 1, 1))

run_single_lm("a ~ b",
              data=data)

}
